latest autoconf/automake update

+++ coders/dng.c
@@ -346,7 +346,7 @@ static void LibRawDataError(void *data,const char *magick_unused(file),
       */
       exception=(ExceptionInfo *) data;
       (void) ThrowMagickException(exception,GetMagickModule(),
-        CorruptImageWarning,"Data corrupted at","`%d'",offset);
+        CorruptImageWarning,"Data corrupted at","`%.20g'",(double) offset);
   }
 }
